+++ src/ofxGgmlMusic.h
@@ -5,5 +5,6 @@
 #include "ofxGgmlMusic/ofxGgmlMusicMidiUtils.h"
 #include "ofxGgmlMusic/ofxGgmlMusicTypes.h"
 #include "ofxGgmlMusic/ofxGgmlMusicGenerationBackend.h"
+#include "ofxGgmlMusic/ofxGgmlMusicExternalGenerationBackend.h"
 #include "ofxGgmlMusic/ofxGgmlMusicProceduralGenerationBackend.h"
 #include "ofxGgmlMusic/ofxGgmlMusicUtils.h"